src/test/java/com/qa/testcases/LoginPageTest.java
package com.qa.testcases;

import org.testng.Assert;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Listeners;
import org.testng.annotations.Test;
import org.apache.log4j.Logger;


import com.aventstack.extentreports.model.Log;
import com.qa.baseclass.BaseClass;
import com.qa.pages.HomePage;
import com.qa.pages.LoginPage;
import com.qa.utilities.CustomListener;

@Listeners(CustomListener.class)
public class LoginPageTest extends BaseClass{
	LoginPage login;
	HomePage homepage;

	
	public LoginPageTest() {
		super();
	}
	
	@BeforeMethod
	public void setUP() {
		initialization();
		log.info("Lauching Chrome Browser");
		login = new LoginPage();
	}
	
	@Test(priority=1)
	public void validateTitle() {
		log.info("****************************** Starting test cases execution  *****************************************");
		log.info("****************************** freeCrmTitleTest *****************************************");
		String title = "";
		try {
		title = login.validateLoginPageTitle();
		log.info("Validating page Title --> " + title);
		Assert.assertEquals(title, "#1 Free CRM software for any business - FreeCRM.co");
		}catch(Exception e) {
			log.error("Login page title is not showing as expected --> " + title);
		}catch(AssertionError e) {
			log.error(e);
		}
		log.info("****************************** ending test case *****************************************");
		log.info("****************************** freeCrmTitleTest *****************************************");
	}
	@Test(priority=2)
	public void validateImage() {
		log.info("****************************** Starting test cases execution  *****************************************");
		log.info("****************************** freeCrmImageTest *****************************************");
		boolean image = login.validateCRMImage();
		log.info("Validating Image --> " + image);
		Assert.assertTrue(image);
		log.warn("Asserted : image");
		log.info("****************************** ending test case *****************************************");
		log.info("****************************** freeCrmImageTest *****************************************");
	}
	@Test(priority=3)
	public void loginTest() {	
		log.info("****************************** Starting test cases execution   *****************************************");
		log.info("****************************** freeCrmLoginTest *****************************************");
		log.info("Entering UserName --> " + propertie.getProperty("UN"));
		String userName = propertie.getProperty("UN");
		log.info("Entering Password --> " + propertie.getProperty("PWD"));
		String passWord = propertie.getProperty("PWD");
		log.info("Click on login");
		homepage = login.login(userName, passWord);
		log.info("Return homepage object --> " + homepage);
		log.info("****************************** ending test case *****************************************");
		log.info("****************************** freeCrmLoginTest *****************************************");
	}
	
	@AfterMethod
	public void tearDown() {
		log.info("Quit the browser");
		driver.quit();
	}
	
	
}
